newsletter shows up in home page of each author fetching all resources and data dynamically
