+++ src/pages/signup.js
@@ -60,6 +60,21 @@ const SignUp = () => {
     return obj
   }
 
+  // create newsletters container
+  const CreateNewslettersContainer = async () => {
+    try {
+      const api = `${process.env.API_HOST}/api/newsletters/create`
+      const response = await axios.post(api, {
+        username: formData.username,
+        pathname: formData.username,
+        emails: [],
+      })
+      console.log(response)
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
   // api handler function for sign in
   const CallAPI = async () => {
     try {
@@ -75,6 +90,8 @@ const SignUp = () => {
           message: 'Registration successfull!',
         })
         document.getElementById('sign_up_form').reset()
+
+        CreateNewslettersContainer()
 
         // sending back to target location on the site
         Router.push('/signin')
@@ -111,68 +128,71 @@ const SignUp = () => {
   return (
     <NavFooter>
       <div className="container section min-h-[80vh]">
-      <div className="max-w-[380px] p-5 rounded-md shadow-xl border m-auto">
-        
-        {/* message showcase according to api responses */}
-        {message ? (
-          <div
-            className={`${
-              message.type ? 'bg-green-400' : 'bg-red-600'
-            } text-white px-2 py-[4px] rounded-md mb-2 text-center`}
-          >
-            {message.message}
+        <div className="max-w-[380px] p-5 rounded-md shadow-xl border m-auto">
+          {/* message showcase according to api responses */}
+          {message ? (
+            <div
+              className={`${
+                message.type ? 'bg-green-400' : 'bg-red-600'
+              } text-white px-2 py-[4px] rounded-md mb-2 text-center`}
+            >
+              {message.message}
+            </div>
+          ) : null}
+          <h3 className="text-[1rem] font-bold text-[#0891B2] text-center">
+            Welcome to 1link
+          </h3>
+
+          {/* sign up form  */}
+          <form id="sign_up_form">
+            <label htmlFor="username">Username</label>
+            <input
+              type="text"
+              name="username"
+              placeholder="Username"
+              onChange={handleOnChange}
+              id="username"
+            />
+            {errorMessage?.username ? (
+              <span>{errorMessage.username}</span>
+            ) : null}
+            <label htmlFor="email">Email</label>
+            <input
+              type="email"
+              name="email"
+              placeholder="Email"
+              onChange={handleOnChange}
+              id="email"
+            />
+            {errorMessage?.email ? <span>{errorMessage.email}</span> : null}
+
+            <label htmlFor="password">Password</label>
+            <input
+              type="password"
+              name="password"
+              placeholder="Password"
+              onChange={handleOnChange}
+              id="password"
+            />
+            {errorMessage?.password ? (
+              <span>{errorMessage.password}</span>
+            ) : null}
+
+            {/* submit button  */}
+            <Button onClick={handleOnSubmit} disable={spinner}>
+              {spinner ? <Spinner /> : null}
+              {spinner ? 'Processing' : 'Sign up'}
+            </Button>
+          </form>
+
+          <div className="flex items-center justify-center gap-3 mt-3">
+            <p>Already have an account?</p>
+            <Link href="/signin" className="text-[#0891B2]">
+              Sign in
+            </Link>
           </div>
-        ) : null}
-        <h3 className="text-[1rem] font-bold text-[#0891B2] text-center">
-          Welcome to 1link
-        </h3>
-
-        {/* sign up form  */}
-        <form id="sign_up_form">
-          <label htmlFor="username">Username</label>
-          <input
-            type="text"
-            name="username"
-            placeholder="Username"
-            onChange={handleOnChange}
-            id="username"
-          />
-          {errorMessage?.username ? <span>{errorMessage.username}</span> : null}
-          <label htmlFor="email">Email</label>
-          <input
-            type="email"
-            name="email"
-            placeholder="Email"
-            onChange={handleOnChange}
-            id="email"
-          />
-          {errorMessage?.email ? <span>{errorMessage.email}</span> : null}
-
-          <label htmlFor="password">Password</label>
-          <input
-            type="password"
-            name="password"
-            placeholder="Password"
-            onChange={handleOnChange}
-            id="password"
-          />
-          {errorMessage?.password ? <span>{errorMessage.password}</span> : null}
-
-          {/* submit button  */}
-          <Button onClick={handleOnSubmit} disable={spinner}>
-            {spinner ? <Spinner /> : null}
-            {spinner ? 'Processing' : 'Sign up'}
-          </Button>
-        </form>
-
-        <div className="flex items-center justify-center gap-3 mt-3">
-          <p>Already have an account?</p>
-          <Link href="/signin" className="text-[#0891B2]">
-            Sign in
-          </Link>
         </div>
       </div>
-    </div>
     </NavFooter>
   )
 }
